fix: arruma teclado na tela de conversas e adicionar contatos
