+++ FamilySafeTalk_Front/App.tsx
@@ -22,7 +22,7 @@ export default function App() {
     <SafeAreaView style={{ flex : 1, paddingTop: 35, marginTop: 2, backgroundColor: '#FFAFCC'}}>
       <View style={styles.container}>
         <StatusBar style="auto" />
-        <Routes />
+        <Conversa />
       </View>
     </SafeAreaView>
   );
