+++ FamilySafeTalk_Front/src/components/contact/index.tsx
@@ -3,6 +3,7 @@ import { ButtonContainer, Container } from "./styles";
 import CardContacts from "../cardContacts";
 
 export default function Contatos() {
+  
   return (
     <Container>
       <ButtonContainer>
