+++ FamilySafeTalk_Front/src/components/topBar/styles.ts
@@ -13,7 +13,7 @@ export const LeftContents = styled(View)`
   flex-direction: row;
   align-items: center;
   justify-content: start;
-  padding: 5% 0 0 2%;
+  padding: 0 0 0 2%;
   width: 60%;
 `;
 
@@ -33,5 +33,7 @@ export const TextUser = styled(Text)`
 export const RightContent = styled(View)`
   width: 20%;
   flex-direction: row;
-  padding: 6% 2% 0 0;
+  align-items: center;
+  justify-content: center;
+  padding: 0 2% 0 0;
 `;
