+++ FamilySafeTalk_Front/src/screens/conversa/index.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { GestureHandlerRootView } from "react-native-gesture-handler";
+import { GestureHandlerRootView, ScrollView } from "react-native-gesture-handler";
 import { Container, Talks } from "./styles";
 import TopBar from "../../components/topBar";
 import BottomBar from "../../components/bottomBar";
@@ -7,7 +7,7 @@ import Conversation from "../../components/conversation";
 
 export default function Conversa() {
   return (
-    <GestureHandlerRootView>
+    <GestureHandlerRootView style={{ flex: 1, backgroundColor: 'white' }}>
       <Container>
         <TopBar />
         <Talks>
