+++ FamilySafeTalk_Front/src/screens/conversa/styles.ts
@@ -4,7 +4,6 @@ import { View, Text, TouchableOpacity, TextInput } from "react-native";
 export const Container = styled(View)`
   flex-grow: 1;
   height: 100%;
-  background-color: white;
 `;
 
 export const Talks = styled(View)`
